Rename `warn.ts` to `console.ts`
