+++ src/components/VPopover/VPopoverContent.vue
@@ -21,7 +21,7 @@
 <script>
 import { defineComponent, toRefs, ref, provide } from '@nuxtjs/composition-api'
 import { usePopoverContent } from '~/composables/use-popover-content'
-import { warn } from '~/utils/warn'
+import { warn } from '~/utils/console'
 
 import { propTypes } from './VPopoverContent.types'
