+++ src/store/index.js
@@ -1,7 +1,7 @@
 import { FETCH_MEDIA_PROVIDERS } from '~/constants/action-types'
 import { PROVIDER } from '~/constants/store-modules'
 
-import { warn } from '~/utils/warn'
+import { warn } from '~/utils/console'
 
 export const actions = {
   async nuxtServerInit({ dispatch }) {
